added sad test to cryptid by name
