+++ spec/graphql/queries/cryptids/cryptid_by_name_spec.rb
@@ -13,6 +13,20 @@ module Queries
         expect(result['data']['cryptidByName']['range']).to be_a String
         expect(result['data']['cryptidByName']['image']).to be_a String
       end
+
+      it 'returns an error if no cryptid by name' do
+        create(:cryptid, name: 'asdawdt')
+
+        result = OkCryptidBeSchema.execute(cryptid_by_name_query).as_json
+        expect(result).to have_key("errors")
+      end
+
+      it 'returns an error if cryptid name is left blank' do
+        create(:cryptid, name: 'asdawdt')
+
+        result = OkCryptidBeSchema.execute(cryptid_by_name_query_blank_name).as_json
+        expect(result).to have_key("errors")
+      end
     end
   end
 end
@@ -21,6 +35,20 @@ def cryptid_by_name_query
   <<~GQL
     {
       cryptidByName(name: "Big Foot") {
+        name
+        description
+        dangerLevel
+        range
+        image
+        }
+    }
+  GQL
+end
+
+def cryptid_by_name_query_blank_name
+  <<~GQL
+    {
+      cryptidByName(name: "") {
         name
         description
         dangerLevel
